fix(bug): wide view mode, TOC when update blocks
